.
- add formats to metadata
- start working on tiled forward pipeline, error atm due to blending. We need independent blending feature to be able to choose whether we blend colour attachments individually, and pipeline creation is error-ing because the 2nd attachment has a format which may not have the COLOUR_ATTACHMENT_BLEND_BIT, even though both formats in the dynamic pipeline creation info do have this bit according to the docs.

+++ examples/tiled-forward/shaders/clustered.frag
@@ -11,7 +11,7 @@ layout(location = 3) in vec4 ClipPos;
 layout(location = 4) in vec4 LastClipPos;
 
 layout(location = 0) out vec4 outColor;
-layout(location = 1) out vec4 outPacking;
+layout(location = 1) out uvec4 outPacking;
 
 layout(set = 0, binding = 0) uniform InstanceUBO {
     mat4 u_model;
